+++ pepper/rc/changelog.md
@@ -15,6 +15,8 @@
 - fix crash when opening and editing the middle of a buffer in a single action (in a macro or through lsp for example)
 - fix `` [aA]` `` was missing
 - added `log` as a `open` command option which is the same as old `scratch` and now `scratch` enables buffer history by default
+- added `file-backed` as a `open` command option which affects open and reopen commands
+- added `to-lowercase` and `to-uppercase` commands
 
 # 0.29.0
 - added diagnostic logging to `spawn` command